+++ student_alexander_keda/src/main/java/org/javaguru/travel/insurance/core/util/AgeCalculator.java
@@ -8,11 +8,11 @@ import java.time.temporal.ChronoUnit;
 @Component
 public class AgeCalculator {
 
-    public long calculate(LocalDate birthDate) {
-        long age = ChronoUnit.YEARS.between(birthDate, LocalDate.now());
+    public int calculate(LocalDate birthDate) {
+        int age = (int) ChronoUnit.YEARS.between(birthDate, LocalDate.now());
         return age >= 0
                 ? age
-                : -1L;
+                : -1;
     }
 
 }
