+++ student_alexander_keda/src/test/resources/schema.sql
@@ -31,8 +31,8 @@ CREATE TABLE country_default_day_rate (
 
 CREATE TABLE IF NOT EXISTS age_coefficient (
 	id BIGINT NOT NULL AUTO_INCREMENT,
-    age_from SMALLINT NOT NULL,
-    age_to SMALLINT NOT NULL,
+    age_from INT NOT NULL,
+    age_to INT NOT NULL,
     coefficient NUMERIC(10,2) NOT NULL,
     PRIMARY KEY (id)
 );
